requirements.txt
pandas
requests
MetaTrader5
python-dateutil